Explicitly instantiating each variation of DataArray<T>
This should cut down on compile times (somewhat).
