+++ Source/SIMPLib/CoreFilters/ImportHDF5Dataset.cpp
@@ -41,6 +41,7 @@
 #include "H5Support/H5ScopedSentinel.h"
 #include "H5Support/H5Utilities.h"
 #include "H5Support/QH5Utilities.h"
+#include "H5Support/QH5Lite.h"
 
 #include "SIMPLib/Common/Constants.h"
 #include "SIMPLib/FilterParameters/AttributeMatrixSelectionFilterParameter.h"
