+++ Source/SIMPLib/DataArrays/DataArray.hpp
@@ -35,27 +35,20 @@
 
 #pragma once
 
+#include <hdf5.h>
+
 // STL Includes
 #include <cassert>
-#include <cstring>
-#include <functional>
-#include <iostream>
-#include <numeric>
-#include <string>
 #include <vector>
+#include <iterator>
+#include <memory>
 
+#include <QtCore/QString>
 #include <QtCore/QTextStream>
 
 #include "SIMPLib/SIMPLib.h"
-
+#include "SIMPLib/Common/Constants.h"
 #include "SIMPLib/DataArrays/IDataArray.h"
-#include "SIMPLib/HDF5/H5DataArrayReader.h"
-#include "SIMPLib/HDF5/H5DataArrayWriter.hpp"
-
-#define mxa_bswap(s, d, t)                                                                                                                                                                             \
-  t[0] = ptr[s];                                                                                                                                                                                       \
-  ptr[s] = ptr[d];                                                                                                                                                                                     \
-  ptr[d] = t[0];
 
 /**
  * @class DataArray
@@ -71,34 +64,23 @@ public:
   using ConstPointer = std::shared_ptr<const Self>;
   using WeakPointer = std::weak_ptr<Self>;
   using ConstWeakPointer = std::weak_ptr<const Self>;
-  static Pointer NullPointer()
-  {
-    return Pointer(static_cast<Self*>(nullptr));
-  }
+
+  static Pointer NullPointer();
 
   /**
    * @brief Returns the name of the class for AbstractMessage
    */
-  QString getNameOfClass() const override
-  {
-    return QString("DataArray<T>");
-  }
+  QString getNameOfClass() const override;
   /**
    * @brief Returns the name of the class for AbstractMessage
    */
-  static QString ClassName()
-  {
-    return QString("DataArray<T>");
-  }
+  static QString ClassName();
 
   /**
    * @brief Returns the version of this class.
    * @return
    */
-  int32_t getClassVersion() const override
-  {
-    return 2;
-  }
+  int32_t getClassVersion() const override;
 
   DataArray(const DataArray&) = default;           // Copy Constructor default Implemented
   DataArray(DataArray&&) = delete;                 // Move Constructor Not Implemented
@@ -118,7 +100,7 @@ public:
   using ContainterType = std::vector<Pointer>;
 
   //========================================= Constructing DataArray Objects =================================
-  DataArray() = default;
+  DataArray();
 
   /**
    * @brief Constructor
@@ -126,13 +108,7 @@ public:
    * @param name The name of the DataArray
    * @param initValue The value to use when initializing each element of the array
    */
-  DataArray(size_t numTuples, const QString& name, T initValue)
-  : IDataArray(name)
-  , m_InitValue(initValue)
-  {
-    resizeTuples(numTuples);
-  }
-
+  DataArray(size_t numTuples, const QString& name, T initValue);
   /**
    * @brief DataArray
    * @param numTuples The number of Tuples in the DataArray
@@ -142,15 +118,7 @@ public:
    *
    * For example if you have a 2D image dimensions of 80(w) x 60(h) then the "cdims" would be [80][60]
    */
-  DataArray(size_t numTuples, const QString& name, comp_dims_type compDims, T initValue)
-  : IDataArray(name)
-  , m_NumTuples(numTuples)
-  , m_CompDims(std::move(compDims))
-  {
-    m_NumComponents = std::accumulate(m_CompDims.begin(), m_CompDims.end(), static_cast<size_t>(1), std::multiplies<>());
-    m_InitValue = initValue;
-    m_Array = resizeAndExtend(m_NumTuples * m_NumComponents);
-  }
+  DataArray(size_t numTuples, const QString& name, comp_dims_type compDims, T initValue);
 
   /**
    * @brief Protected Constructor
@@ -160,31 +128,9 @@ public:
    * @param initValue The value to use when initializing each element of the array
    * @param allocate Will all the memory be allocated at time of construction
    */
-  DataArray(size_t numTuples, const QString& name, comp_dims_type compDims, T initValue, bool allocate)
-  : IDataArray(name)
-  , m_NumTuples(numTuples)
-  , m_CompDims(std::move(compDims))
-  {
-    m_NumComponents = std::accumulate(m_CompDims.begin(), m_CompDims.end(), static_cast<size_t>(1), std::multiplies<>());
-    m_InitValue = initValue;
-    if(allocate)
-    {
-      resizeTuples(numTuples);
-    }
-    else
-    {
-      m_Size = m_NumTuples * m_NumComponents;
-      m_MaxId = (m_Size > 0) ? m_Size - 1 : m_Size;
-    }
-#if 0
-      MUD_FLAP_0 = MUD_FLAP_1 = MUD_FLAP_2 = MUD_FLAP_3 = MUD_FLAP_4 = MUD_FLAP_5 = 0xABABABABABABABABul;
-#endif
-  }
+  DataArray(size_t numTuples, const QString& name, comp_dims_type compDims, T initValue, bool allocate);
 
-  ~DataArray() override
-  {
-    clear();
-  }
+  ~DataArray() override;
 
   //========================================= Static Constructing DataArray Objects =================================
   /**
@@ -194,26 +140,7 @@ public:
    * @param allocate Will all the memory be allocated at time of construction
    * @return Std::Shared_Ptr wrapping an instance of DataArrayTemplate<T>
    */
-  static Pointer CreateArray(size_t numTuples, const QString& name, bool allocate)
-  {
-    if(name.isEmpty())
-    {
-      return NullPointer();
-    }
-    comp_dims_type cDims = {1};
-    auto d = new DataArray<T>(numTuples, name, cDims, static_cast<T>(0), allocate);
-    if(allocate)
-    {
-      if(d->allocate() < 0)
-      {
-        // Could not allocate enough memory, reset the pointer to null and return
-        delete d;
-        return DataArray<T>::NullPointer();
-      }
-    }
-    Pointer ptr(d);
-    return ptr;
-  }
+  static Pointer CreateArray(size_t numTuples, const QString& name, bool allocate);
 
   /**
    * @brief Static constructor
@@ -224,30 +151,7 @@ public:
    * @param allocate Will all the memory be allocated at time of construction
    * @return Std::Shared_Ptr wrapping an instance of DataArrayTemplate<T>
    */
-  static Pointer CreateArray(size_t numTuples, int rank, const size_t* dims, const QString& name, bool allocate)
-  {
-    if(name.isEmpty())
-    {
-      return NullPointer();
-    }
-    comp_dims_type cDims(rank);
-    for(int i = 0; i < rank; i++)
-    {
-      cDims[i] = dims[i];
-    }
-    auto d = new DataArray<T>(numTuples, name, cDims, static_cast<T>(0), allocate);
-    if(allocate)
-    {
-      if(d->allocate() < 0)
-      {
-        // Could not allocate enough memory, reset the pointer to null and return
-        delete d;
-        return DataArray<T>::NullPointer();
-      }
-    }
-    Pointer ptr(d);
-    return ptr;
-  }
+  static Pointer CreateArray(size_t numTuples, int rank, const size_t* dims, const QString& name, bool allocate);
 
   /**
    * @brief Static constructor
@@ -257,25 +161,7 @@ public:
    * @param allocate Will all the memory be allocated at time of construction
    * @return Std::Shared_Ptr wrapping an instance of DataArrayTemplate<T>
    */
-  static Pointer CreateArray(size_t numTuples, const comp_dims_type& compDims, const QString& name, bool allocate)
-  {
-    if(name.isEmpty())
-    {
-      return NullPointer();
-    }
-    DataArray<T>* d = new DataArray<T>(numTuples, name, compDims, static_cast<T>(0), allocate);
-    if(allocate)
-    {
-      if(d->allocate() < 0)
-      {
-        // Could not allocate enough memory, reset the pointer to null and return
-        delete d;
-        return DataArray<T>::NullPointer();
-      }
-    }
-    Pointer ptr(d);
-    return ptr;
-  }
+  static Pointer CreateArray(size_t numTuples, const comp_dims_type& compDims, const QString& name, bool allocate);
 
   /**
    * @brief Static constructor
@@ -286,28 +172,7 @@ public:
    * @param allocate Will all the memory be allocated at time of construction
    * @return Std::Shared_Ptr wrapping an instance of DataArrayTemplate<T>
    */
-  static Pointer CreateArray(const comp_dims_type& tupleDims, const comp_dims_type& compDims, const QString& name, bool allocate)
-  {
-    if(name.isEmpty())
-    {
-      return NullPointer();
-    }
-
-    size_t numTuples = std::accumulate(tupleDims.begin(), tupleDims.end(), static_cast<size_t>(1), std::multiplies<>());
-
-    auto d = new DataArray<T>(numTuples, name, compDims, static_cast<T>(0), allocate);
-    if(allocate)
-    {
-      if(d->allocate() < 0)
-      {
-        // Could not allocate enough memory, reset the pointer to null and return
-        delete d;
-        return DataArray<T>::NullPointer();
-      }
-    }
-    Pointer ptr(d);
-    return ptr;
-  }
+  static Pointer CreateArray(const comp_dims_type& tupleDims, const comp_dims_type& compDims, const QString& name, bool allocate);
 
   //========================================= Instance Constructing DataArray Objects =================================
   /**
@@ -319,11 +184,7 @@ public:
    * @param allocate Will all the memory be allocated at time of construction
    * @return
    */
-  IDataArray::Pointer createNewArray(size_t numTuples, int rank, const size_t* compDims, const QString& name, bool allocate) const override
-  {
-    IDataArray::Pointer p = DataArray<T>::CreateArray(numTuples, rank, compDims, name, allocate);
-    return p;
-  }
+  IDataArray::Pointer createNewArray(size_t numTuples, int rank, const size_t* compDims, const QString& name, bool allocate) const override;
 
   /**
    * @brief createNewArray
@@ -333,11 +194,7 @@ public:
    * @param allocate Will all the memory be allocated at time of construction
    * @return Std::Shared_Ptr wrapping an instance of DataArrayTemplate<T>
    */
-  IDataArray::Pointer createNewArray(size_t numTuples, const comp_dims_type& compDims, const QString& name, bool allocate) const override
-  {
-    IDataArray::Pointer p = DataArray<T>::CreateArray(numTuples, compDims, name, allocate);
-    return p;
-  }
+  IDataArray::Pointer createNewArray(size_t numTuples, const comp_dims_type& compDims, const QString& name, bool allocate) const override;
 
   /**
    * @brief Static Method to create a DataArray from a QVector through a deep copy of the data
@@ -346,16 +203,7 @@ public:
    * @param name The name of the array
    * @return Std::Shared_Ptr wrapping an instance of DataArrayTemplate<T>
    */
-  static Pointer FromQVector(QVector<T>& vec, const QString& name)
-  {
-
-    Pointer p = CreateArray(vec.size(), name, true);
-    if(nullptr != p.get())
-    {
-      std::memcpy(p->getPointer(0), vec.data(), vec.size() * sizeof(T));
-    }
-    return p;
-  }
+  static Pointer FromQVector(QVector<T>& vec, const QString& name);
 
   /**
    * @brief Static Method to create a DataArray from a std::vector through a deep copy of the data
@@ -364,16 +212,7 @@ public:
    * @param name The name of the array
    * @return Std::Shared_Ptr wrapping an instance of DataArrayTemplate<T>
    */
-  static Pointer FromStdVector(std::vector<T>& vec, const QString& name)
-  {
-    comp_dims_type cDims = {1};
-    Pointer p = CreateArray(vec.size(), cDims, name, true);
-    if(nullptr != p.get())
-    {
-      std::memcpy(p->getPointer(0), &(vec.front()), vec.size() * sizeof(T));
-    }
-    return p;
-  }
+  static Pointer FromStdVector(std::vector<T>& vec, const QString& name);
 
   /**
    * @brief FromPointer Creates a DataArray<T> object with a <b>DEEP COPY</b> of the data
@@ -382,15 +221,7 @@ public:
    * @param name
    * @return
    */
-  static Pointer CopyFromPointer(T* data, size_t size, const QString& name)
-  {
-    Pointer p = CreateArray(size, name, true);
-    if(nullptr != p.get())
-    {
-      std::memcpy(p->getPointer(0), data, size * sizeof(T));
-    }
-    return p;
-  }
+  static Pointer CopyFromPointer(T* data, size_t size, const QString& name);
 
   /**
    * @brief WrapPointer Creates a DataArray<T> object that references the pointer. The original caller can
@@ -403,22 +234,7 @@ public:
    * @param ownsData
    * @return
    */
-  static Pointer WrapPointer(T* data, size_t numTuples, const comp_dims_type& compDims, const QString& name, bool ownsData)
-  {
-    // Allocate on the heap
-    auto d = new DataArray(numTuples, name, compDims, static_cast<T>(0), false);
-    // Wrap that heap pointer with a shared_pointer to make it reference counted
-    Pointer p(d);
-
-    p->m_Array = data;        // Now set the internal array to the raw pointer
-    p->m_OwnsData = ownsData; // Set who owns the data, i.e., who is going to "free" the memory
-    if(nullptr != data)
-    {
-      p->m_IsAllocated = true;
-    }
-
-    return p;
-  }
+  static Pointer WrapPointer(T* data, size_t numTuples, const comp_dims_type& compDims, const QString& name, bool ownsData);
 
   //========================================= Begin API =================================
 
@@ -427,23 +243,7 @@ public:
    * @param forceNoAllocate
    * @return
    */
-  IDataArray::Pointer deepCopy(bool forceNoAllocate = false) const override
-  {
-    bool allocate = m_IsAllocated;
-    if(forceNoAllocate)
-    {
-      allocate = false;
-    }
-    IDataArray::Pointer daCopy = createNewArray(getNumberOfTuples(), getComponentDimensions(), getName(), allocate);
-    if(m_IsAllocated && !forceNoAllocate)
-    {
-      T* src = getPointer(0);
-      void* dest = daCopy->getVoidPointer(0);
-      size_t totalBytes = (getNumberOfTuples() * getNumberOfComponents() * sizeof(T));
-      std::memcpy(dest, src, totalBytes);
-    }
-    return daCopy;
-  }
+  IDataArray::Pointer deepCopy(bool forceNoAllocate = false) const override;
 
   /**
    * @brief GetTypeName Returns a string representation of the type of data that is stored by this class. This
@@ -451,133 +251,14 @@ public:
    * @return
    */
 
-  SIMPL::NumericTypes::Type getType() const
-  {
-    T value = static_cast<T>(0x00);
-    if(typeid(value) == typeid(int8_t))
-    {
-      return SIMPL::NumericTypes::Type::Int8;
-    }
-    if(typeid(value) == typeid(uint8_t))
-    {
-      return SIMPL::NumericTypes::Type::UInt8;
-    }
-
-    if(typeid(value) == typeid(int16_t))
-    {
-      return SIMPL::NumericTypes::Type::Int16;
-    }
-    if(typeid(value) == typeid(uint16_t))
-    {
-      return SIMPL::NumericTypes::Type::UInt16;
-    }
-
-    if(typeid(value) == typeid(int32_t))
-    {
-      return SIMPL::NumericTypes::Type::Int32;
-    }
-    if(typeid(value) == typeid(uint32_t))
-    {
-      return SIMPL::NumericTypes::Type::UInt32;
-    }
-
-    if(typeid(value) == typeid(int64_t))
-    {
-      return SIMPL::NumericTypes::Type::Int64;
-    }
-    if(typeid(value) == typeid(uint64_t))
-    {
-      return SIMPL::NumericTypes::Type::UInt64;
-    }
-
-    if(typeid(value) == typeid(float))
-    {
-      return SIMPL::NumericTypes::Type::Float;
-    }
-    if(typeid(value) == typeid(double))
-    {
-      return SIMPL::NumericTypes::Type::Double;
-    }
-
-    if(typeid(value) == typeid(bool))
-    {
-      return SIMPL::NumericTypes::Type::Bool;
-    }
-
-    return SIMPL::NumericTypes::Type::UnknownNumType;
-  }
+  SIMPL::NumericTypes::Type getType() const;
 
   /**
    * @brief GetTypeName Returns a string representation of the type of data that is stored by this class. This
    * can be a primitive like char, float, int or the name of a class.
    * @return
    */
-  void getXdmfTypeAndSize(QString& xdmfTypeName, int& precision) const override
-  {
-    T value = static_cast<T>(0x00);
-    xdmfTypeName = "UNKNOWN";
-    precision = 0;
-    if(typeid(value) == typeid(int8_t))
-    {
-      xdmfTypeName = "Char";
-      precision = 1;
-    }
-    if(typeid(value) == typeid(uint8_t))
-    {
-      xdmfTypeName = "UChar";
-      precision = 1;
-    }
-
-    if(typeid(value) == typeid(int16_t))
-    {
-      xdmfTypeName = "Int";
-      precision = 2;
-    }
-    if(typeid(value) == typeid(uint16_t))
-    {
-      xdmfTypeName = "UInt";
-      precision = 2;
-    }
-
-    if(typeid(value) == typeid(int32_t))
-    {
-      xdmfTypeName = "Int";
-      precision = 4;
-    }
-    if(typeid(value) == typeid(uint32_t))
-    {
-      xdmfTypeName = "UInt";
-      precision = 4;
-    }
-
-    if(typeid(value) == typeid(int64_t))
-    {
-      xdmfTypeName = "Int";
-      precision = 8;
-    }
-    if(typeid(value) == typeid(uint64_t))
-    {
-      xdmfTypeName = "UInt";
-      precision = 8;
-    }
-
-    if(typeid(value) == typeid(float))
-    {
-      xdmfTypeName = "Float";
-      precision = 4;
-    }
-    if(typeid(value) == typeid(double))
-    {
-      xdmfTypeName = "Float";
-      precision = 8;
-    }
-
-    if(typeid(value) == typeid(bool))
-    {
-      xdmfTypeName = "uchar";
-      precision = 1;
-    }
-  }
+  void getXdmfTypeAndSize(QString& xdmfTypeName, int& precision) const override;
   // This line must be here, because we are overloading the copyData pure virtual function in IDataArray.
   // This is required so that other classes can call this version of copyData from the subclasses.
   using IDataArray::copyFromArray;
@@ -601,161 +282,52 @@ public:
    * @param sourceArray
    * @return
    */
-  bool copyFromArray(size_t destTupleOffset, IDataArray::Pointer sourceArray, size_t srcTupleOffset, size_t totalSrcTuples) override
-  {
-    if(!m_IsAllocated)
-    {
-      return false;
-    }
-    if(nullptr == m_Array)
-    {
-      return false;
-    }
-    if(destTupleOffset > m_MaxId)
-    {
-      return false;
-    }
-    if(!sourceArray->isAllocated())
-    {
-      return false;
-    }
-    Self* source = dynamic_cast<Self*>(sourceArray.get());
-    if(nullptr == source->getPointer(0))
-    {
-      return false;
-    }
-
-    if(sourceArray->getNumberOfComponents() != getNumberOfComponents())
-    {
-      return false;
-    }
-
-    if(srcTupleOffset + totalSrcTuples > sourceArray->getNumberOfTuples())
-    {
-      return false;
-    }
-
-    if(totalSrcTuples * sourceArray->getNumberOfComponents() + destTupleOffset * getNumberOfComponents() > m_Size)
-    {
-      return false;
-    }
-
-    size_t elementStart = destTupleOffset * getNumberOfComponents();
-    size_t totalBytes = (totalSrcTuples * sourceArray->getNumberOfComponents()) * sizeof(T);
-    std::memcpy(m_Array + elementStart, source->getPointer(srcTupleOffset * sourceArray->getNumberOfComponents()), totalBytes);
-    return true;
-  }
+  bool copyFromArray(size_t destTupleOffset, IDataArray::Pointer sourceArray, size_t srcTupleOffset, size_t totalSrcTuples) override;
 
   /**
    * @brief copyIntoArray
    * @param dest
    */
-  bool copyIntoArray(Pointer dest) const
-  {
-    if(m_IsAllocated && dest->isAllocated() && m_Array && dest->getPointer(0))
-    {
-      size_t totalBytes = m_Size * sizeof(T);
-      std::memcpy(dest->getPointer(0), m_Array, totalBytes);
-      return true;
-    }
-    return false;
-  }
+  bool copyIntoArray(Pointer dest) const;
 
   /**
    * @brief isAllocated
    * @return
    */
-  bool isAllocated() const override
-  {
-    return m_IsAllocated;
-  }
-
+  bool isAllocated() const override;
   /**
    * @brief Gives this array a human readable name
    * @param name The name of this array
    */
-  virtual void setInitValue(T initValue)
-  {
-    m_InitValue = initValue;
-  }
+  virtual void setInitValue(T initValue);
 
   /**
    * @brief Makes this class responsible for freeing the memory
    */
-  void takeOwnership() override
-  {
-    m_OwnsData = true;
-  }
+  void takeOwnership() override;
 
   /**
    * @brief This class will NOT free the memory associated with the internal pointer.
    * This can be useful if the user wishes to keep the data around after this
    * class goes out of scope.
    */
-  void releaseOwnership() override
-  {
-    m_OwnsData = false;
-  }
+  void releaseOwnership() override;
 
   /**
    * @brief Allocates the memory needed for this class
    * @return 1 on success, -1 on failure
    */
-  virtual int32_t allocate()
-  {
-    if((nullptr != m_Array) && (true == m_OwnsData))
-    {
-      deallocate();
-    }
-    m_Array = nullptr;
-    m_OwnsData = true;
-    m_IsAllocated = false;
-    if(m_Size == 0)
-    {
-      clear();
-      return 1;
-    }
-
-    size_t newSize = m_Size;
-    m_Array = new T[newSize]();
-    if(!m_Array)
-    {
-      qDebug() << "Unable to allocate " << newSize << " elements of size " << sizeof(T) << " bytes. ";
-      return -1;
-    }
-    m_Size = newSize;
-    m_IsAllocated = true;
-
-    return 1;
-  }
+  virtual int32_t allocate();
 
   /**
    * @brief Sets all the values to zero.
    */
-  void initializeWithZeros() override
-  {
-    if(!m_IsAllocated || nullptr == m_Array)
-    {
-      return;
-    }
-    size_t typeSize = sizeof(T);
-    ::memset(m_Array, 0, m_Size * typeSize);
-  }
+  void initializeWithZeros() override;
 
   /**
    * @brief Sets all the values to value.
    */
-  virtual void initializeWithValue(T initValue, size_t offset = 0)
-  {
-    if(!m_IsAllocated || nullptr == m_Array)
-    {
-      return;
-    }
-    for(size_t i = offset; i < m_Size; i++)
-    {
-      m_Array[i] = initValue;
-    }
-  }
+  virtual void initializeWithValue(T initValue, size_t offset = 0);
 
   /**
    * @brief Removes Tuples from the m_Array. If the size of the vector is Zero nothing is done. If the size of the
@@ -765,116 +337,7 @@ public:
    * @param idxs The indices to remove
    * @return error code.
    */
-  int eraseTuples(comp_dims_type& idxs) override
-  {
-    int err = 0;
-
-    // If nothing is to be erased just return
-    if(idxs.empty())
-    {
-      return 0;
-    }
-    auto idxs_size = static_cast<size_t>(idxs.size());
-    if(idxs_size >= getNumberOfTuples())
-    {
-      resizeTuples(0);
-      return 0;
-    }
-
-    // Sanity Check the Indices in the vector to make sure we are not trying to remove any indices that are
-    // off the end of the array and return an error code.
-    for(size_t& idx : idxs)
-    {
-      if(idx * m_NumComponents > m_MaxId)
-      {
-        return -100;
-      }
-    }
-
-    // Calculate the new size of the array to copy into
-    size_t newSize = (getNumberOfTuples() - idxs.size()) * m_NumComponents;
-
-    // Create a new m_Array to copy into
-    T* newArray = new T[newSize]();
-
-    // Splat AB across the array so we know if we are copying the values or not
-    ::memset(newArray, 0xAB, newSize * sizeof(T));
-
-    // Keep the current Destination Pointer
-    T* currentDest = newArray;
-    size_t j = 0;
-    int k = 0;
-    // Find the first chunk to copy by walking the idxs array until we get an
-    // index that is NOT a continuous increment from the start
-    for(k = 0; k < idxs.size(); ++k)
-    {
-      if(j == idxs[k])
-      {
-        ++j;
-      }
-      else
-      {
-        break;
-      }
-    }
-
-    if(k == idxs.size()) // Only front elements are being dropped
-    {
-      T* currentSrc = m_Array + (j * m_NumComponents);
-      std::memcpy(currentDest, currentSrc, (getNumberOfTuples() - idxs.size()) * m_NumComponents * sizeof(T));
-      deallocate(); // We are done copying - delete the current m_Array
-      m_Size = newSize;
-      m_Array = newArray;
-      m_OwnsData = true;
-      m_MaxId = newSize - 1;
-      m_IsAllocated = true;
-      return 0;
-    }
-
-    comp_dims_type srcIdx(idxs.size() + 1);
-    comp_dims_type destIdx(idxs.size() + 1);
-    comp_dims_type copyElements(idxs.size() + 1);
-    srcIdx[0] = 0;
-    destIdx[0] = 0;
-    copyElements[0] = (idxs[0] - 0) * m_NumComponents;
-
-    for(int i = 1; i < srcIdx.size(); ++i)
-    {
-      srcIdx[i] = (idxs[i - 1] + 1) * m_NumComponents;
-
-      if(i < srcIdx.size() - 1)
-      {
-        copyElements[i] = (idxs[i] - idxs[i - 1] - 1) * m_NumComponents;
-      }
-      else
-      {
-        copyElements[i] = (getNumberOfTuples() - idxs[i - 1] - 1) * m_NumComponents;
-      }
-      destIdx[i] = copyElements[i - 1] + destIdx[i - 1];
-    }
-
-    // Copy the data
-    for(int i = 0; i < srcIdx.size(); ++i)
-    {
-      currentDest = newArray + destIdx[i];
-      T* currentSrc = m_Array + srcIdx[i];
-      size_t bytes = copyElements[i] * sizeof(T);
-      std::memcpy(currentDest, currentSrc, bytes);
-    }
-
-    // We are done copying - delete the current m_Array
-    deallocate();
-
-    // Allocation was successful.  Save it.
-    m_Size = newSize;
-    m_Array = newArray;
-    // This object has now allocated its memory and owns it.
-    m_OwnsData = true;
-    m_IsAllocated = true;
-    m_MaxId = newSize - 1;
-
-    return err;
-  }
+  int eraseTuples(comp_dims_type& idxs) override;
 
   /**
    * @brief
@@ -882,19 +345,7 @@ public:
    * @param newPos
    * @return
    */
-  int copyTuple(size_t currentPos, size_t newPos) override
-  {
-    size_t max = ((m_MaxId + 1) / m_NumComponents);
-    if(currentPos >= max || newPos >= max)
-    {
-      return -1;
-    }
-    T* src = m_Array + (currentPos * m_NumComponents);
-    T* dest = m_Array + (newPos * m_NumComponents);
-    size_t bytes = sizeof(T) * m_NumComponents;
-    std::memcpy(dest, src, bytes);
-    return 0;
-  }
+  int copyTuple(size_t currentPos, size_t newPos) override;
 
   /**
    * @brief Returns the number of bytes that make up the data type.
@@ -903,46 +354,31 @@ public:
    * 4 = 32 bit integer/Float
    * 8 = 64 bit integer/Double
    */
-  size_t getTypeSize() const override
-  {
-    return sizeof(T);
-  }
+  size_t getTypeSize() const override;
 
   /**
    * @brief Returns the number of elements in the internal array.
    */
-  size_t getNumberOfTuples() const override
-  {
-    return m_NumTuples;
-  }
+  size_t getNumberOfTuples() const override;
 
   /**
    * @brief Returns the total number of elements that make up this array. Equal to NumTuples * NumComponents
    */
-  size_t getSize() const override
-  {
-    return m_Size;
-  }
+  size_t getSize() const override;
 
   /**
    * @brief Returns the dimensions for the data residing at each Tuple. For example if you have a simple Scalar value
    * at each tuple then this will return a single element QVector. If you have a 1x3 array (like EUler Angles) then
    * this will return a 3 Element QVector.
    */
-  comp_dims_type getComponentDimensions() const override
-  {
-    return m_CompDims;
-  }
+  comp_dims_type getComponentDimensions() const override;
 
   /**
    * @brief Returns the number component values at each Tuple location. For example if you have a
    * 3 element component (vector) then this will be 3. If you are storing a small image of size 80x60
    * at each Tuple (like EBSD Kikuchi patterns) then the result would be 4800.
    */
-  int getNumberOfComponents() const override
-  {
-    return m_NumComponents;
-  }
+  int getNumberOfComponents() const override;
 
   /**
    * @brief Returns a void pointer pointing to the index of the array. nullptr
@@ -951,41 +387,19 @@ public:
    * @param i The index to have the returned pointer pointing to.
    * @return Void Pointer. Possibly nullptr.
    */
-  void* getVoidPointer(size_t i) override
-  {
-    if(i >= m_Size)
-    {
-      return nullptr;
-    }
-
-    return (void*)(&(m_Array[i]));
-  }
+  void* getVoidPointer(size_t i) override;
 
   /**
    * @brief Returns a list of the contents of DataArray (For Python Binding)
    * @return std::list. Possibly empty
    */
-  std::list<T> getArray() const
-  {
-    return std::list<T>(m_Array, m_Array + (m_Size * sizeof(T)) / sizeof(T));
-  }
+  std::list<T> getArray() const;
 
   /**
    * @brief Sets the contents of the array to the list (For Python Binding)
    * @param std::list. New array contents
    */
-  void setArray(std::list<T> newArray)
-  {
-    if(newArray.size() != m_Size)
-    {
-      return;
-    }
-    int i = 0;
-    for(auto elem : newArray)
-    {
-      m_Array[i++] = elem;
-    }
-  }
+  void setArray(std::list<T> newArray);
 
   /**
    * @brief Returns the pointer to a specific index into the array. No checks are made
@@ -994,61 +408,25 @@ public:
    * @param i The index to return the pointer to.
    * @return The pointer to the index
    */
-  virtual T* getPointer(size_t i) const
-  {
-#ifndef NDEBUG
-    if(m_Size > 0)
-    {
-      Q_ASSERT(i < m_Size);
-    }
-#endif
-    return (T*)(&(m_Array[i]));
-  }
+  virtual T* getPointer(size_t i) const;
 
   /**
    * @brief Returns the value for a given index
    * @param i The index to return the value at
    * @return The value at index i
    */
-  virtual T getValue(size_t i) const
-  {
-#ifndef NDEBUG
-    if(m_Size > 0)
-    {
-      Q_ASSERT(i < m_Size);
-    }
-#endif
-    return m_Array[i];
-  }
+  virtual T getValue(size_t i) const;
 
   /**
    * @brief Sets a specific value in the array
    * @param i The index of the value to set
    * @param value The new value to be set at the specified index
    */
-  void setValue(size_t i, T value)
-  {
-#ifndef NDEBUG
-    if(m_Size > 0)
-    {
-      Q_ASSERT(i < m_Size);
-    }
-#endif
-    m_Array[i] = value;
-  }
+  void setValue(size_t i, T value);
 
   //----------------------------------------------------------------------------
   // These can be overridden for more efficiency
-  T getComponent(size_t i, int j) const
-  {
-#ifndef NDEBUG
-    if(m_Size > 0)
-    {
-      Q_ASSERT(i * m_NumComponents + j < m_Size);
-    }
-#endif
-    return m_Array[i * m_NumComponents + j];
-  }
+  T getComponent(size_t i, int j) const;
 
   /**
    * @brief Sets a specific component of the Tuple located at i
@@ -1056,105 +434,41 @@ public:
    * @param j The Component index into the Tuple
    * @param c The value to set
    */
-  void setComponent(size_t i, int j, T c)
-  {
-#ifndef NDEBUG
-    if(m_Size > 0)
-    {
-      Q_ASSERT(i * m_NumComponents + j < m_Size);
-    }
-#endif
-    m_Array[i * m_NumComponents + j] = c;
-  }
+  void setComponent(size_t i, int j, T c);
 
   /**
    * @brief setTuple
    * @param tupleIndex
    * @param data
    */
-  void setTuple(size_t tupleIndex, T* data)
-  {
-#ifndef NDEBUG
-    if(m_Size > 0)
-    {
-      Q_ASSERT(tupleIndex * m_NumComponents + (m_NumComponents - 1) < m_Size);
-    }
-#endif
-    std::memcpy(getTuplePointer(tupleIndex), data, m_NumComponents * sizeof(T));
-  }
+  void setTuple(size_t tupleIndex, T* data);
 
   /**
    * @brief setTuple
    * @param tupleIndex
    * @param data
    */
-  void setTuple(size_t tupleIndex, const std::vector<T>& data)
-  {
-#ifndef NDEBUG
-    if(m_Size > 0)
-    {
-      Q_ASSERT(tupleIndex * m_NumComponents + (m_NumComponents - 1) < m_Size);
-    }
-#endif
-    std::memcpy(getTuplePointer(tupleIndex), data.data(), m_NumComponents * sizeof(T));
-  }
+  void setTuple(size_t tupleIndex, const std::vector<T>& data);
 
   /**
    * @brief Splats the same value c across all values in the Tuple
    * @param i The index of the Tuple
    * @param c The value to splat across all components in the tuple
    */
-  void initializeTuple(size_t i, void* p) override
-  {
-    if(!m_IsAllocated)
-    {
-      return;
-    }
-#ifndef NDEBUG
-    if(m_Size > 0)
-    {
-      Q_ASSERT(i * m_NumComponents < m_Size);
-    }
-#endif
-    if(nullptr == p)
-    {
-      return;
-    }
-    T* c = reinterpret_cast<T*>(p);
-    for(size_t j = 0; j < m_NumComponents; ++j)
-    {
-      m_Array[i * m_NumComponents + j] = *c;
-    }
-  }
+  void initializeTuple(size_t i, void* p) override;
 
   /**
    * @brief getTuplePointer Returns the pointer to a specific tuple
    * @param tupleIndex The index of tuple
    */
-  T* getTuplePointer(size_t tupleIndex) const
-  {
-#ifndef NDEBUG
-    if(m_Size > 0)
-    {
-      Q_ASSERT(tupleIndex * m_NumComponents < m_Size);
-    }
-#endif
-    return m_Array + (tupleIndex * m_NumComponents);
-  }
+  T* getTuplePointer(size_t tupleIndex) const;
 
   /**
    * @brief resize
    * @param numTuples
    * @return
    */
-  void resizeTuples(size_t numTuples) override
-  {
-    T* ptr = resizeAndExtend(numTuples * m_NumComponents);
-    if(nullptr != ptr)
-    {
-      m_NumTuples = numTuples;
-    }
-  }
+  void resizeTuples(size_t numTuples) override;
 
   /**
    * @brief printTuple
@@ -1162,29 +476,7 @@ public:
    * @param i
    * @param delimiter
    */
-  void printTuple(QTextStream& out, size_t i, char delimiter = ',') const override
-  {
-    int precision = out.realNumberPrecision();
-    T value = static_cast<T>(0x00);
-    if(typeid(value) == typeid(float))
-    {
-      out.setRealNumberPrecision(8);
-    }
-    if(typeid(value) == typeid(double))
-    {
-      out.setRealNumberPrecision(16);
-    }
-
-    for(size_t j = 0; j < m_NumComponents; ++j)
-    {
-      if(j != 0)
-      {
-        out << delimiter;
-      }
-      out << m_Array[i * m_NumComponents + j];
-    }
-    out.setRealNumberPrecision(precision);
-  }
+  void printTuple(QTextStream& out, size_t i, char delimiter = ',') const override;
 
   /**
    * @brief printComponent
@@ -1192,10 +484,7 @@ public:
    * @param i
    * @param j
    */
-  void printComponent(QTextStream& out, size_t i, int j) const override
-  {
-    out << m_Array[i * m_NumComponents + j];
-  }
+  void printComponent(QTextStream& out, size_t i, int j) const override;
 
   /**
    * @brief Returns the HDF Type for a given primitive value.
@@ -1203,180 +492,20 @@ public:
    * from
    * @return The HDF5 native type for the value
    */
-  QString getFullNameOfClass() const
-  {
-    QString theType = getTypeAsString();
-    theType = "DataArray<" + theType + ">";
-    return theType;
-  }
+  QString getFullNameOfClass() const;
 
   /**
    * @brief getTypeAsString
    * @return
    */
-  QString getTypeAsString() const override
-  {
-    T value = static_cast<T>(0);
-    if(typeid(value) == typeid(float))
-    {
-      return "float";
-    }
-    if(typeid(value) == typeid(double))
-    {
-      return "double";
-    }
-
-    if(typeid(value) == typeid(int8_t))
-    {
-      return "int8_t";
-    }
-    if(typeid(value) == typeid(uint8_t))
-    {
-      return "uint8_t";
-    }
-#if CMP_TYPE_CHAR_IS_SIGNED
-    if(typeid(value) == typeid(char))
-    {
-      return "char";
-    }
-#else
-    if(typeid(value) == typeid(char))
-    {
-      return "char";
-    }
-#endif
-    if(typeid(value) == typeid(signed char))
-    {
-      return "signed char";
-    }
-    if(typeid(value) == typeid(unsigned char))
-    {
-      return "unsigned char";
-    }
-
-    if(typeid(value) == typeid(int16_t))
-    {
-      return "int16_t";
-    }
-    if(typeid(value) == typeid(short))
-    {
-      return "short";
-    }
-    if(typeid(value) == typeid(signed short))
-    {
-      return "signed short";
-    }
-    if(typeid(value) == typeid(uint16_t))
-    {
-      return "uint16_t";
-    }
-    if(typeid(value) == typeid(unsigned short))
-    {
-      return "unsigned short";
-    }
-
-    if(typeid(value) == typeid(int32_t))
-    {
-      return "int32_t";
-    }
-    if(typeid(value) == typeid(uint32_t))
-    {
-      return "uint32_t";
-    }
-#if(CMP_SIZEOF_INT == 4)
-    if(typeid(value) == typeid(int))
-    {
-      return "int";
-    }
-    if(typeid(value) == typeid(signed int))
-    {
-      return "signed int";
-    }
-    if(typeid(value) == typeid(unsigned int))
-    {
-      return "unsigned int";
-    }
-#endif
-
-    if(typeid(value) == typeid(int64_t))
-    {
-      return "int64_t";
-    }
-    if(typeid(value) == typeid(uint64_t))
-    {
-      return "uint64_t";
-    }
-
-#if(CMP_SIZEOF_LONG == 4)
-    if(typeid(value) == typeid(long int))
-    {
-      return "long int";
-    }
-    if(typeid(value) == typeid(signed long int))
-    {
-      return "signed long int";
-    }
-    if(typeid(value) == typeid(unsigned long int))
-    {
-      return "unsigned long int";
-    }
-#elif(CMP_SIZEOF_LONG == 8)
-    if(typeid(value) == typeid(long int))
-    {
-      return "long int";
-    }
-    if(typeid(value) == typeid(signed long int))
-    {
-      return "signed long int";
-    }
-    if(typeid(value) == typeid(unsigned long int))
-    {
-      return "unsigned long int";
-    }
-#endif
-
-#if(CMP_SIZEOF_LONG_LONG == 8)
-    if(typeid(value) == typeid(long long int))
-    {
-      return "long long int";
-    }
-    if(typeid(value) == typeid(signed long long int))
-    {
-      return "signed long long int";
-    }
-    if(typeid(value) == typeid(unsigned long long int))
-    {
-      return "unsigned long long int";
-    }
-#endif
-
-    if(typeid(value) == typeid(bool))
-    {
-      return "bool";
-    }
-
-    // qDebug()  << "Error: HDFTypeForPrimitive - Unknown Type: " << (typeid(value).name()) ;
-    const char* name = typeid(value).name();
-    if(nullptr != name && name[0] == 'l')
-    {
-      qDebug() << "You are using 'long int' as a type which is not 32/64 bit safe. Suggest you use one of the H5SupportTypes defined in <Common/H5SupportTypes.h> such as int32_t or uint32_t.";
-    }
-    return "UnknownType";
-  }
+  QString getTypeAsString() const override;
 
   /**
    *
    * @param parentId
    * @return
    */
-  int writeH5Data(hid_t parentId, comp_dims_type tDims) const override
-  {
-    if(m_Array == nullptr)
-    {
-      return -85648;
-    }
-    return H5DataArrayWriter::writeDataArray<Self>(parentId, this, tDims);
-  }
+  int writeH5Data(hid_t parentId, comp_dims_type tDims) const override;
 
   /**
    * @brief writeXdmfAttribute
@@ -1384,170 +513,26 @@ public:
    * @param volDims
    * @return
    */
-  int writeXdmfAttribute(QTextStream& out, int64_t* volDims, const QString& hdfFileName, const QString& groupPath, const QString& label) const override
-  {
-    if(m_Array == nullptr)
-    {
-      return -85648;
-    }
-    QString dimStr;
-    int precision = 0;
-    QString xdmfTypeName;
-    getXdmfTypeAndSize(xdmfTypeName, precision);
-    if(0 == precision)
-    {
-      out << "<!-- " << getName() << " has unknown type or unsupported type or precision for XDMF to understand"
-          << " -->"
-          << "\n";
-      return -100;
-    }
-
-    int numComp = getNumberOfComponents();
-    out << "    <Attribute Name=\"" << getName() << label << "\" ";
-    if(numComp == 1)
-    {
-      out << "AttributeType=\"Scalar\" ";
-      dimStr = QString("%1 %2 %3 ").arg(volDims[2]).arg(volDims[1]).arg(volDims[0]);
-    }
-    else if(numComp == 6)
-    {
-      out << "AttributeType=\"Tensor6\" ";
-      dimStr = QString("%1 %2 %3 %4 ").arg(volDims[2]).arg(volDims[1]).arg(volDims[0]).arg(numComp);
-    }
-    else if(numComp == 9)
-    {
-      out << "AttributeType=\"Tensor\" ";
-      dimStr = QString("%1 %2 %3 %4 ").arg(volDims[2]).arg(volDims[1]).arg(volDims[0]).arg(numComp);
-    }
-    else
-    {
-      out << "AttributeType=\"Vector\" ";
-      dimStr = QString("%1 %2 %3 %4 ").arg(volDims[2]).arg(volDims[1]).arg(volDims[0]).arg(numComp);
-    }
-    out << "Center=\"Cell\">\n";
-    // Open the <DataItem> Tag
-    out << R"(      <DataItem Format="HDF" Dimensions=")" << dimStr << R"(" )";
-    out << "NumberType=\"" << xdmfTypeName << "\" "
-        << "Precision=\"" << precision << "\" >\n";
-
-    out << "        " << hdfFileName << groupPath << "/" << getName() << "\n";
-    out << "      </DataItem>"
-        << "\n";
-    out << "    </Attribute>"
-        << "\n";
-    return 1;
-  }
+  int writeXdmfAttribute(QTextStream& out, int64_t* volDims, const QString& hdfFileName, const QString& groupPath, const QString& label) const override;
 
   /**
    * @brief getInfoString
    * @return Returns a formatted string that contains general infomation about
    * the instance of the object.
    */
-  QString getInfoString(SIMPL::InfoStringFormat format) const override
-  {
-
-    QLocale usa(QLocale::English, QLocale::UnitedStates);
-
-    QString info;
-    QTextStream ss(&info);
-    if(format == SIMPL::HtmlFormat)
-    {
-      ss << "<html><head></head>\n";
-      ss << "<body>\n";
-      ss << "<table cellpadding=\"4\" cellspacing=\"0\" border=\"0\">\n";
-      ss << "<tbody>\n";
-      ss << "<tr bgcolor=\"#FFFCEA\"><th colspan=2>Attribute Array Info</th></tr>";
-
-      ss << R"(<tr bgcolor="#E9E7D6"><th align="right">Name:</th><td>)" << getName() << "</td></tr>";
-
-      ss << R"(<tr bgcolor="#FFFCEA"><th align="right">Type:</th><td> DataArray&lt;)" << getTypeAsString() << "&gt;</td></tr>";
-      QString numStr = usa.toString(static_cast<qlonglong>(getNumberOfTuples()));
-      ss << R"(<tr bgcolor="#FFFCEA"><th align="right">Number of Tuples:</th><td>)" << numStr << "</td></tr>";
-
-      QString compDimStr = "(";
-      for(int i = 0; i < m_CompDims.size(); i++)
-      {
-        compDimStr = compDimStr + QString::number(m_CompDims[i]);
-        if(i < m_CompDims.size() - 1)
-        {
-          compDimStr = compDimStr + QString(", ");
-        }
-      }
-      compDimStr = compDimStr + ")";
-      ss << R"(<tr bgcolor="#FFFCEA"><th align="right">Component Dimensions:</th><td>)" << compDimStr << "</td></tr>";
-      numStr = usa.toString(static_cast<qlonglong>(m_Size));
-      ss << R"(<tr bgcolor="#FFFCEA"><th align="right">Total Elements:</th><td>)" << numStr << "</td></tr>";
-      numStr = usa.toString(static_cast<qlonglong>(m_Size * sizeof(T)));
-      ss << R"(<tr bgcolor="#FFFCEA"><th align="right">Total Memory Required:</th><td>)" << numStr << "</td></tr>";
-      ss << "</tbody></table>\n";
-      ss << "</body></html>";
-    }
-    else
-    {
-    }
-    return info;
-  }
+  QString getInfoString(SIMPL::InfoStringFormat format) const override;
 
   /**
    * @brief
    * @param parentId
    * @return
    */
-  int readH5Data(hid_t parentId) override
-  {
-    int err = 0;
-
-    resizeTuples(0);
-    IDataArray::Pointer p = H5DataArrayReader::ReadIDataArray(parentId, getName());
-    if(p.get() == nullptr)
-    {
-      return -1;
-    }
-    m_Array = reinterpret_cast<T*>(p->getVoidPointer(0));
-    m_Size = p->getSize();
-    m_OwnsData = true;
-    m_MaxId = (m_Size == 0) ? 0 : m_Size - 1;
-    m_IsAllocated = true;
-    setName(p->getName());
-    m_NumTuples = p->getNumberOfTuples();
-    m_CompDims = p->getComponentDimensions();
-    m_NumComponents = p->getNumberOfComponents();
-
-    // Tell the intermediate DataArray to release ownership of the data as we are going to be responsible
-    // for deleting the memory
-    p->releaseOwnership();
-    return err;
-  }
+  int readH5Data(hid_t parentId) override;
 
   /**
    * @brief
    */
-  virtual void byteSwapElements()
-  {
-    char* ptr = (char*)(m_Array);
-    char t[8];
-    size_t size = getTypeSize();
-    for(uint64_t var = 0; var < m_Size; ++var)
-    {
-      if(sizeof(T) == 2)
-      {
-        mxa_bswap(0, 1, t);
-      }
-      else if(sizeof(T) == 4)
-      {
-        mxa_bswap(0, 3, t);
-        mxa_bswap(1, 2, t);
-      }
-      else if(sizeof(T) == 8)
-      {
-        mxa_bswap(0, 7, t);
-        mxa_bswap(1, 6, t);
-        mxa_bswap(2, 5, t);
-        mxa_bswap(3, 4, t);
-      }
-      ptr += size; // increment the pointer
-    }
-  }
+  virtual void byteSwapElements();
 
   //========================================= STL INTERFACE COMPATIBILITY =================================
 
@@ -1571,8 +556,9 @@ public:
       ptr_ = ptr_ + num_comps_;
       return *this;
     } // PREFIX
-    self_type operator++(int junk)
+    self_type operator++(int ununsed)
     {
+      std::ignore = ununsed;
       self_type i = *this;
       ptr_ = ptr_ + num_comps_;
       return i;
@@ -1623,8 +609,9 @@ public:
       ptr_ = ptr_ + num_comps_;
       return *this;
     } // PREFIX
-    self_type operator++(int junk)
+    self_type operator++(int ununsed)
     {
+      std::ignore = ununsed;
       self_type i = *this;
       ptr_ = ptr_ + num_comps_;
       return i;
@@ -1672,6 +659,7 @@ public:
     iterator(pointer ptr, size_type ununsed)
     : ptr_(ptr)
     {
+      std::ignore = ununsed;
     }
 
     self_type operator++()
@@ -1679,8 +667,9 @@ public:
       ptr_++;
       return *this;
     } // PREFIX
-    self_type operator++(int junk)
+    self_type operator++(int ununsed)
     {
+      std::ignore = ununsed;
       self_type i = *this;
       ptr_++;
       return i;
@@ -1767,34 +756,42 @@ public:
   };
 
   // ######### Iterators #########
-
+  /**
+   *
+   */
   template <typename IteratorType> IteratorType begin()
   {
     return IteratorType(m_Array, m_NumComponents);
   }
-  iterator begin()
-  {
-    return iterator(m_Array);
-  }
+
+  /**
+   * @brief begin
+   * @return
+   */
+  iterator begin();
 
   template <typename IteratorType> IteratorType end()
   {
     return IteratorType(m_Array + m_Size, m_NumComponents);
   }
-  iterator end()
-  {
-    return iterator(m_Array + m_Size);
-  }
 
-  const_iterator begin() const
-  {
-    return const_iterator(m_Array);
-  }
+  /**
+   * @brief end
+   * @return
+   */
+  iterator end();
 
-  const_iterator end() const
-  {
-    return const_iterator(m_Array + m_Size);
-  }
+  /**
+   * @brief begin
+   * @return
+   */
+  const_iterator begin() const;
+
+  /**
+   * @brief end
+   * @return
+   */
+  const_iterator end() const;
 
   // rbegin
   // rend
@@ -1805,28 +802,16 @@ public:
 
   // ######### Capacity #########
 
-  size_type size() const
-  {
-    return m_Size;
-  }
+  size_type size() const;
 
-  size_type max_size() const
-  {
-    return m_Size;
-  }
+  size_type max_size() const;
   //  void resize(size_type n) override
   //  {
   //    resizeAndExtend(n);
   //  }
   // void resize (size_type n, const value_type& val);
-  size_type capacity() const noexcept
-  {
-    return m_Size;
-  }
-  bool empty() const noexcept
-  {
-    return (m_Size == 0);
-  }
+  size_type capacity() const noexcept;
+  bool empty() const noexcept;
   // reserve()
   // shrink_to_fit()
 
@@ -1889,7 +874,8 @@ public:
    * @brief In the range version (1), the new contents are elements constructed from each of the elements in the range
    * between first and last, in the same order.
    */
-  template <class InputIterator> void assign(InputIterator first, InputIterator last) // range (1)
+  template <class InputIterator>
+  void assign(InputIterator first, InputIterator last) // range (1)
   {
     size_type size = last - first;
     resizeAndExtend(size);
@@ -1906,50 +892,29 @@ public:
    * @param n
    * @param val
    */
-  void assign(size_type n, const value_type& val) // fill (2)
-  {
-    resizeAndExtend(n);
-    for(size_t i = 0; i < n; i++)
-    {
-      m_Array[i] = val;
-    }
-  }
+  void assign(size_type n, const value_type& val);
 
   /**
    * @brief In the initializer list version (3), the new contents are copies of the values passed as initializer list, in the same order.
    * @param il
    */
-  void assign(std::initializer_list<value_type> il) //  initializer list (3)
-  {
-    assign(il.begin(), il.end());
-  }
+  void assign(std::initializer_list<value_type> il);
 
   /**
    * @brief push_back
    * @param val
    */
-  void push_back(const value_type& val)
-  {
-    resizeAndExtend(m_Size + 1);
-    m_Array[m_MaxId] = val;
-  }
+  void push_back(const value_type& val);
   /**
    * @brief push_back
    * @param val
    */
-  void push_back(value_type&& val)
-  {
-    resizeAndExtend(m_Size + 1);
-    m_Array[m_MaxId] = val;
-  }
+  void push_back(value_type&& val);
 
   /**
    * @brief pop_back
    */
-  void pop_back()
-  {
-    resizeAndExtend(m_Size - 1);
-  }
+  void pop_back();
   // insert
   // iterator erase (const_iterator position)
   // iterator erase (const_iterator first, const_iterator last);
@@ -1958,19 +923,7 @@ public:
   /**
    * @brief Removes all elements from the array (which are destroyed), leaving the container with a size of 0.
    */
-  void clear()
-  {
-    if(nullptr != m_Array && m_OwnsData)
-    {
-      deallocate();
-    }
-    m_Array = nullptr;
-    m_Size = 0;
-    m_OwnsData = true;
-    m_MaxId = 0;
-    m_IsAllocated = false;
-    m_NumTuples = 0;
-  }
+  void clear();
   // emplace
   // emplace_back
 
@@ -1980,7 +933,8 @@ public:
    * @param range2
    * @return
    */
-  template <typename Range1, typename Range2> bool equal(Range1 const& range1, Range2 const& range2)
+  template <typename Range1, typename Range2>
+  bool equal(Range1 const& range1, Range2 const& range2)
   {
     if(range1.size() != range2.size())
     {
@@ -1997,137 +951,21 @@ protected:
   /**
    * @brief deallocates the memory block
    */
-  void deallocate()
-  {
-    // We are going to splat 0xABABAB across the first value of the array as a debugging aid
-    auto cptr = reinterpret_cast<unsigned char*>(m_Array);
-    if(nullptr != cptr)
-    {
-      if(m_Size > 0)
-      {
-        if(sizeof(T) >= 1)
-        {
-          cptr[0] = 0xAB;
-        }
-        if(sizeof(T) >= 2)
-        {
-          cptr[1] = 0xAB;
-        }
-        if(sizeof(T) >= 4)
-        {
-          cptr[2] = 0xAB;
-          cptr[3] = 0xAB;
-        }
-        if(sizeof(T) >= 8)
-        {
-          cptr[4] = 0xAB;
-          cptr[5] = 0xAB;
-          cptr[6] = 0xAB;
-          cptr[7] = 0xAB;
-        }
-      }
-    }
-#if 0
-      if (MUD_FLAP_0 != 0xABABABABABABABABul
-          || MUD_FLAP_1 != 0xABABABABABABABABul
-          || MUD_FLAP_2 != 0xABABABABABABABABul
-          || MUD_FLAP_3 != 0xABABABABABABABABul
-          || MUD_FLAP_4 != 0xABABABABABABABABul
-          || MUD_FLAP_5 != 0xABABABABABABABABul)
-      {
-        Q_ASSERT(false);
-      }
-#endif
-    delete[](m_Array);
-
-    m_Array = nullptr;
-    m_IsAllocated = false;
-  }
+  void deallocate();
 
   /**
    * @brief Resizes the internal array
    * @param size The new size of the internal array
    * @return 1 on success, 0 on failure
    */
-  int32_t resizeTotalElements(size_t size) override
-  {
-    // std::cout << "DataArray::resizeTotalElements(" << size << ")" << std::endl;
-    if(size == 0)
-    {
-      clear();
-      return 1;
-    }
-    T* ptr = resizeAndExtend(size);
-    if(nullptr != ptr)
-    {
-      return 1;
-    }
-    return 0;
-  }
+  int32_t resizeTotalElements(size_t size) override;
 
   /**
    * @brief resizes the internal array to be 'size' elements in length
    * @param size
    * @return Pointer to the internal array
    */
-  virtual T* resizeAndExtend(size_t size)
-  {
-    T* newArray = nullptr;
-    size_t newSize = 0;
-    size_t oldSize = 0;
-
-    if(size == m_Size) // Requested size is equal to current size.  Do nothing.
-    {
-      return m_Array;
-    }
-    newSize = size;
-    oldSize = m_Size;
-
-    // Wipe out the array completely if new size is zero.
-    if(newSize == 0)
-    {
-      clear();
-      return m_Array;
-    }
-
-    newArray = new T[newSize]();
-    if(!newArray)
-    {
-      qDebug() << "Unable to allocate " << newSize << " elements of size " << sizeof(T) << " bytes. ";
-      return nullptr;
-    }
-
-    // Copy the data from the old array.
-    if(m_Array != nullptr)
-    {
-      std::memcpy(newArray, m_Array, (newSize < m_Size ? newSize : m_Size) * sizeof(T));
-    }
-
-    // Allocate a new array if we DO NOT own the current array
-    if((nullptr != m_Array) && m_OwnsData)
-    {
-      // Free the old array
-      deallocate();
-    }
-
-    // Allocation was successful.  Save it.
-    m_Size = newSize;
-    m_Array = newArray;
-
-    // This object has now allocated its memory and owns it.
-    m_OwnsData = true;
-
-    m_MaxId = newSize - 1;
-    m_IsAllocated = true;
-
-    // Initialize the new tuples if newSize is larger than old size
-    if(newSize > oldSize)
-    {
-      initializeWithValue(m_InitValue, oldSize);
-    }
-
-    return m_Array;
-  }
+  virtual T* resizeAndExtend(size_t size);
 
 private:
   T* m_Array = nullptr;
@@ -2142,11 +980,37 @@ private:
 };
 
 // -----------------------------------------------------------------------------
-//
+// These are specialized for bool type as std::vector<bool> uses bits instead of bytes
+template <>
+typename DataArray<bool>::Pointer DataArray<bool>::FromStdVector(std::vector<bool>& vec, const QString& name);
+
+template <>
+void DataArray<bool>::setTuple(size_t tupleIndex, const std::vector<bool>& data);
+
 // -----------------------------------------------------------------------------
+// Declare our extern templates
+extern template class DataArray<bool>;
+
+extern template class DataArray<char>;
+extern template class DataArray<unsigned char>;
+
+extern template class DataArray<int8_t>;
+extern template class DataArray<uint8_t>;
+extern template class DataArray<int16_t>;
+extern template class DataArray<uint16_t>;
+extern template class DataArray<int32_t>;
+extern template class DataArray<uint32_t>;
+extern template class DataArray<int64_t>;
+extern template class DataArray<uint64_t>;
+
+extern template class DataArray<float>;
+extern template class DataArray<double>;
+
+extern template class DataArray<size_t>;
 
 using BoolArrayType = DataArray<bool>;
 
+using CharArrayType = DataArray<char>;
 using UCharArrayType = DataArray<unsigned char>;
 
 using Int8ArrayType = DataArray<int8_t>;
