+++ Source/SIMPLib/DataContainers/DataContainerBundle.cpp
@@ -35,13 +35,13 @@
 
 #include "DataContainerBundle.h"
 
-#include "H5Support/QH5Utilities.h"
-#include "H5Support/H5ScopedSentinel.h"
-
 #include <QtCore/QDebug>
 
-#include "SIMPLib/DataArrays/StringDataArray.h"
+#include "H5Support/QH5Utilities.h"
+#include "H5Support/QH5Lite.h"
+#include "H5Support/H5ScopedSentinel.h"
 
+#include "SIMPLib/DataArrays/StringDataArray.h"
 #include "SIMPLib/DataContainers/DataArrayPath.h"
 #include "SIMPLib/DataContainers/DataContainer.h"
