+++ Source/SIMPLib/Utilities/SIMPLH5DataReader.cpp
@@ -36,15 +36,15 @@
 
 #include <QtCore/QDebug>
 
-#include "SIMPLib/Common/IObserver.h"
+#include "H5Support/QH5Lite.h"
+#include "H5Support/QH5Utilities.h"
+#include "H5Support/H5ScopedSentinel.h"
 
+#include "SIMPLib/Common/IObserver.h"
 #include "SIMPLib/DataContainers/DataContainerArray.h"
 #include "SIMPLib/DataContainers/DataContainerBundle.h"
 #include "SIMPLib/Utilities/SIMPLH5DataReaderRequirements.h"
 #include "SIMPLib/DataContainers/DataContainer.h"
-
-#include "H5Support/QH5Utilities.h"
-#include "H5Support/H5ScopedSentinel.h"
 
 const QString Title = "HDF5 Read Error";
