+++ Source/SVWidgetsLib/FilterParameterWidgets/ImportHDF5DatasetWidget.cpp
@@ -42,6 +42,7 @@
 
 #include <QtGui/QKeyEvent>
 #include <QtGui/QPainter>
+#include <QtCore/QTextStream>
 
 #include <QtWidgets/QFileDialog>
 #include <QtWidgets/QMenu>
@@ -49,11 +50,9 @@
 #include "H5Support/H5Lite.h"
 #include "H5Support/H5Utilities.h"
 #include "H5Support/QH5Utilities.h"
-
-#include <QtCore/QTextStream>
+#include "H5Support/QH5Lite.h"
 
 #include "SIMPLib/CoreFilters/ImportHDF5Dataset.h"
-
 #include "SIMPLib/FilterParameters/ImportHDF5DatasetFilterParameter.h"
 #include "SIMPLib/Utilities/SIMPLDataPathValidator.h"
 #include "SIMPLib/DataContainers/DataContainerArray.h"
